requirements.txt
agora-token-builder==1.0.0
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
colorama==0.4.6
constantly==15.1.0
coverage==7.3.0
cryptography==41.0.3
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.4
django-environ==0.11.2
django-sslserver==0.22
factory-boy==3.3.0
Faker==19.3.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.7
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
setuptools==68.2.2
six==1.16.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
zope.interface==6.0
